refactor and fixed integration tests
