+++ autokeras/__init__.py
@@ -1,30 +1,30 @@
 from autokeras.auto_model import AutoModel
-from autokeras.hypermodel.base import Block
-from autokeras.hypermodel.base import Head
-from autokeras.hypermodel.base import Node
-from autokeras.hypermodel.block import ConvBlock
-from autokeras.hypermodel.block import DenseBlock
-from autokeras.hypermodel.block import Embedding
-from autokeras.hypermodel.block import FeatureEncoding
-from autokeras.hypermodel.block import ImageAugmentation
-from autokeras.hypermodel.block import ImageBlock
-from autokeras.hypermodel.block import Merge
-from autokeras.hypermodel.block import Normalization
-from autokeras.hypermodel.block import ResNetBlock
-from autokeras.hypermodel.block import RNNBlock
-from autokeras.hypermodel.block import SpatialReduction
-from autokeras.hypermodel.block import StructuredDataBlock
-from autokeras.hypermodel.block import TemporalReduction
-from autokeras.hypermodel.block import TextBlock
-from autokeras.hypermodel.block import TextToIntSequence
-from autokeras.hypermodel.block import TextToNgramVector
-from autokeras.hypermodel.block import XceptionBlock
-from autokeras.hypermodel.head import ClassificationHead
-from autokeras.hypermodel.head import RegressionHead
-from autokeras.hypermodel.node import ImageInput
-from autokeras.hypermodel.node import Input
-from autokeras.hypermodel.node import StructuredDataInput
-from autokeras.hypermodel.node import TextInput
+from autokeras.engine.block import Block
+from autokeras.engine.block import Node
+from autokeras.hypermodels.basic import ConvBlock
+from autokeras.hypermodels.basic import DenseBlock
+from autokeras.hypermodels.basic import Embedding
+from autokeras.hypermodels.basic import ResNetBlock
+from autokeras.hypermodels.basic import RNNBlock
+from autokeras.hypermodels.basic import XceptionBlock
+from autokeras.hypermodels.head import ClassificationHead
+from autokeras.hypermodels.head import Head
+from autokeras.hypermodels.head import RegressionHead
+from autokeras.hypermodels.node import ImageInput
+from autokeras.hypermodels.node import Input
+from autokeras.hypermodels.node import StructuredDataInput
+from autokeras.hypermodels.node import TextInput
+from autokeras.hypermodels.preprocessing import FeatureEncoding
+from autokeras.hypermodels.preprocessing import ImageAugmentation
+from autokeras.hypermodels.preprocessing import Normalization
+from autokeras.hypermodels.preprocessing import TextToIntSequence
+from autokeras.hypermodels.preprocessing import TextToNgramVector
+from autokeras.hypermodels.reduction import Merge
+from autokeras.hypermodels.reduction import SpatialReduction
+from autokeras.hypermodels.reduction import TemporalReduction
+from autokeras.hypermodels.wrapper import ImageBlock
+from autokeras.hypermodels.wrapper import StructuredDataBlock
+from autokeras.hypermodels.wrapper import TextBlock
 from autokeras.task import ImageClassifier
 from autokeras.task import ImageRegressor
 from autokeras.task import StructuredDataClassifier
