+++ autokeras/hypermodels/graph.py
@@ -1,12 +1,108 @@
+import queue
+
 import kerastuner
 import tensorflow as tf
 from tensorflow.python.util import nest
 
-from autokeras.hypermodel import base
-from autokeras.hypermodel import compiler
+from autokeras.engine import picklable
+from autokeras.hypermodels import basic
+from autokeras.hypermodels import head as head_module
+from autokeras.hypermodels import node as node_module
+from autokeras.hypermodels import preprocessing
+from autokeras.hypermodels import reduction
+from autokeras.hypermodels import wrapper
 
 
-class Graph(kerastuner.HyperModel, base.Picklable):
+def embedding_max_features(embedding_block):
+    """Fetch the max_features value for embedding block from TextToIntSequence."""
+    if embedding_block.max_features:
+        return
+    input_node = embedding_block.inputs[0]
+    while True:
+        if not input_node.in_blocks:
+            raise ValueError('If Embedding block is not using with '
+                             'TextToIntSequence, max_features must be '
+                             'specified.')
+        block = input_node.in_blocks[0]
+        if isinstance(block, preprocessing.TextToIntSequence):
+            embedding_block.max_features = block.max_tokens
+            return
+        input_node = block.inputs[0]
+
+
+def fetch_heads(source_block):
+    """Get the downstream head blocks for a given block in the network.
+
+    # Arguments
+        source_block: Block. The source block for the search for heads.
+
+    # Returns
+        A list of Head instances.
+    """
+    heads = []
+    visited_blocks = set()
+    visited_blocks.add(source_block)
+    q = queue.Queue()
+    q.put(source_block)
+    while not q.empty():
+        block = q.get()
+        if isinstance(block, head_module.Head):
+            heads.append(block)
+        for output_node in block.outputs:
+            for next_block in output_node.out_blocks:
+                if next_block not in visited_blocks:
+                    visited_blocks.add(next_block)
+                    q.put(next_block)
+    return heads
+
+
+def feature_encoding_input(block):
+    """Fetch the column_types and column_names.
+
+    The values are fetched for FeatureEncoding from StructuredDataInput.
+    """
+    if not isinstance(block.inputs[0], node_module.StructuredDataInput):
+        raise TypeError('FeatureEncoding block can only be used '
+                        'with StructuredDataInput.')
+    block.column_types = block.inputs[0].column_types
+    block.column_names = block.inputs[0].column_names
+
+
+def structured_data_block_heads(structured_data_block):
+    structured_data_block.num_heads = len(fetch_heads(structured_data_block))
+
+
+# Compile the graph.
+COMPILE_FUNCTIONS = {
+    basic.Embedding: [embedding_max_features],
+    wrapper.StructuredDataBlock: [structured_data_block_heads,
+                                  feature_encoding_input],
+    preprocessing.FeatureEncoding: [feature_encoding_input],
+}
+
+ALL_CLASSES = {
+    **vars(node_module),
+    **vars(head_module),
+    **vars(wrapper),
+    **vars(preprocessing),
+    **vars(basic),
+    **vars(reduction),
+}
+
+
+def serialize(obj):
+    return tf.keras.utils.serialize_keras_object(obj)
+
+
+def deserialize(config, custom_objects=None):
+    return tf.keras.utils.deserialize_keras_object(
+        config,
+        module_objects={**ALL_CLASSES},
+        custom_objects=custom_objects,
+        printable_module_name='graph')
+
+
+class Graph(kerastuner.HyperModel, picklable.Picklable):
     """A graph consists of connected Blocks, or Heads.
 
     # Arguments
@@ -29,16 +125,11 @@ class Graph(kerastuner.HyperModel, base.Picklable):
             self._build_network()
         self.override_hps = override_hps or []
 
-    def compile(self, func):
-        """Share the information between blocks by calling functions in compiler.
-
-        # Arguments
-            func: A dictionary. The keys are the block classes. The values are
-                corresponding compile functions.
-        """
+    def compile(self):
+        """Share the information between blocks."""
         for block in self.blocks:
-            if block.__class__ in func:
-                func[block.__class__](block)
+            for func in COMPILE_FUNCTIONS.get(block.__class__, []):
+                func(block)
 
     def _register_hps(self, hp):
         """Register the override HyperParameters for current HyperParameters."""
@@ -153,8 +244,8 @@ class Graph(kerastuner.HyperModel, base.Picklable):
         raise ValueError('Cannot find block named {name}.'.format(name=name))
 
     def get_config(self):
-        blocks = [compiler.serialize(block) for block in self.blocks]
-        nodes = {str(self._node_to_id[node]): compiler.serialize(node)
+        blocks = [serialize(block) for block in self.blocks]
+        nodes = {str(self._node_to_id[node]): serialize(node)
                  for node in self.inputs}
         override_hps = [tf.keras.utils.serialize_keras_object(hp)
                         for hp in self.override_hps]
@@ -180,8 +271,8 @@ class Graph(kerastuner.HyperModel, base.Picklable):
 
     @classmethod
     def from_config(cls, config):
-        blocks = [compiler.deserialize(block) for block in config['blocks']]
-        nodes = {int(node_id): compiler.deserialize(node)
+        blocks = [deserialize(block) for block in config['blocks']]
+        nodes = {int(node_id): deserialize(node)
                  for node_id, node in config['nodes'].items()}
         override_hps = [kerastuner.engine.hyperparameters.deserialize(config)
                         for config in config['override_hps']]
@@ -217,7 +308,7 @@ class Graph(kerastuner.HyperModel, base.Picklable):
     def build(self, hp):
         """Build the HyperModel into a Keras Model."""
         self._register_hps(hp)
-        self.compile(compiler.COMPILE_FUNCTIONS)
+        self.compile()
         real_nodes = {}
         for input_node in self.inputs:
             node_id = self._node_to_id[input_node]
@@ -241,7 +332,7 @@ class Graph(kerastuner.HyperModel, base.Picklable):
         metrics = {}
         for output_node in self.outputs:
             block = output_node.in_blocks[0]
-            if isinstance(block, base.Head):
+            if isinstance(block, head_module.Head):
                 metrics[block.name] = block.metrics
         return metrics
 
@@ -249,7 +340,7 @@ class Graph(kerastuner.HyperModel, base.Picklable):
         loss = {}
         for output_node in self.outputs:
             block = output_node.in_blocks[0]
-            if isinstance(block, base.Head):
+            if isinstance(block, head_module.Head):
                 loss[block.name] = block.loss
         return loss
