+++ autokeras/keras_layers.py
@@ -121,3 +121,16 @@ class FeatureEncodingCombiner(Combiner):
 
     def deserialize(self, encoded_accumulator):
         pass
+
+
+class Sigmoid(tf.keras.layers.Layer):
+    """Sigmoid activation function."""
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def call(self, inputs):
+        return tf.keras.activations.sigmoid(inputs)
+
+    def compute_output_shape(self, input_shape):
+        return input_shape
