+++ autokeras/tuners/hyperband.py
@@ -1,0 +1,8 @@
+import kerastuner
+
+from autokeras.engine import tuner as tuner_module
+
+
+class Hyperband(tuner_module.AutoTuner, kerastuner.Hyperband):
+    """KerasTuner Hyperband with preprocessing layer tuning."""
+    pass
